chore: rename load test files to v0.3
