+++ crates/load-test/src/tasks.rs
@@ -1,1 +1,1 @@
-pub mod v02;
+pub mod v03;
